Add weather processor. Process current weather for all cities every hour.

+++ WeatherProrok.Logic/Models/CurrentWeatherModel.cs
@@ -7,8 +7,8 @@ namespace WeatherProrok.Logic.Models
 {
     public class CurrentWeatherModel
     {
-        public float Temp { get; set; }
-        public float Humidity { get; set; }
+        public int Temp { get; set; }
+        public int Humidity { get; set; }
         public Cloudity Cloudity { get; set; }
         public Precipitation Precipitation { get; set; }
         public DateTime CurrentDateTime { get; set; }
